+++ app/api/APIResponse.ts
@@ -7,14 +7,14 @@ export function resError<T>(errorMessage: string): APIData<T> {
 	};
 }
 
-export function resUnauthorized(): APIData<{}> {
+export function resUnauthorized<D>(): APIData<D> {
 	return {
 		success: false,
 		errorMessage: "You are not authorized to access this resource!"
 	};
 }
 
-export function resOkData<D extends object>(data: D): APIData<D> {
+export function resOkData<D>(data: D): APIData<D> {
 	return {
 		success: true,
 		...data
@@ -27,20 +27,15 @@ export function resOk(): APIData<{}> {
 	};
 }
 
-export async function resAttemptAsync<R>(fun: () => Promise<R>)
-{
-	try
-	{
+export async function resAttemptAsync<D extends Record<string, any>>(fun: () => Promise<APIData<D>>): Promise<APIData<D>> {
+	try {
 		return await fun();
 	}
-	catch (error)
-	{
-		if (error instanceof Error)
-		{
+	catch (error) {
+		if (error instanceof Error) {
 			return resError(error.message);
 		}
-		else if (error instanceof object)
-		{
+		else if (error instanceof object) {
 			return resError(error.toString())
 		}
 		return resError(error as string);